Move observe functions into new file
